SISG_AUDITORIA/requirements.txt
python-docx>=1.1.0
reportlab>=4.0.0
openpyxl>=3.1.0
Flask>=3.0.0
# Preparacao para WSGI em Linux/container (troca futura de servidor).
gunicorn>=22.0.0; platform_system != "Windows"
# Driver Postgres (fase de transicao para persistencia relacional).
psycopg[binary]>=3.2.0
bcrypt>=4.1.0
PyJWT>=2.8.0
